+++ packages/agents-hosting-express/package.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://json.schemastore.org/package.json",
-  "name": "@microsoft/agents-hosting",
+  "name": "@microsoft/agents-hosting-express",
   "version": "0.1.0",
   "homepage": "https://github.com/microsoft/Agents-for-js",
   "repository": {
@@ -19,12 +19,8 @@
   "main": "dist/src/index.js",
   "types": "dist/src/index.d.ts",
   "dependencies": {
-    "@azure/msal-node": "^3.5.1",
-    "@microsoft/agents-activity": "file:../agents-activity",
-    "axios": "^1.8.2",
-    "debug": "^4.3.7",
-    "jsonwebtoken": "^9.0.2",
-    "jwks-rsa": "^3.2.0"
+    "@microsoft/agents-activity-hosting": "file:../agents-hosting",
+    "express": "^5.10"
   },
   "license": "MIT",
   "files": [
@@ -41,5 +37,8 @@
   },
   "engines": {
     "node": ">=18.0.0"
+  },
+  "devDependencies": {
+    "@types/express": "^5.0.1"
   }
 }
