+++ packages/agents-hosting-express/tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "extends": "../../tsconfig.json",
+    "compilerOptions": {
+      "rootDirs": [ "src/", "test/"],
+      "outDir": "dist",
+      "paths": {
+        "@microsoft/*": ["./packages/*"]
+      }
+    }
+  }
